refacto: move the main into the cmd
